refactor: split parts of the business card
